Fix bad foreign key constraints

+++ db/mathvarstat.sql
@@ -2,7 +2,7 @@
 -- Used by the math search module to analyse the variables in the equations.
 --
 CREATE TABLE /*_*/mathvarstat (
-  varstat_id smallint(6) NOT NULL AUTO_INCREMENT,
+  varstat_id INT(6) NOT NULL AUTO_INCREMENT,
   varstat_featurename varchar(10) NOT NULL,
   varstat_featuretype varchar(10) NOT NULL,
   varstat_featurecount int(11) NOT NULL,
